requirements.txt
aio-pika==6.4.1
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiohttp-requests==0.1.2
aiohttp-security==0.4.0
aiohttp-session==2.9.0
aiojobs==0.2.2
aiormq==3.2.0
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
brotlipy==0.7.0
cachetools==4.0.0
cchardet==2.1.5
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Faker==4.0.2
httplib2==0.17.0
idna==2.8
Jinja2==2.11.1
MarkupSafe==1.1.1
motor==2.1.0
multidict==4.6.1
oauth2client==4.1.3
pamqp==2.3.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.0.0
pycparser==2.19
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==4.0
six==1.14.0
text-unidecode==1.3
tonyg-rfc3339==0.1
trafaret==2.0.2
uritemplate==3.0.1
urllib3==1.25.8
yarl==1.4.1
